Zombie movement and attack animation initiates one Zombie at a time
Moved Zombie movement and attack animation logic inside single for loop
so that attack and movement for zombie only occurs for one at a time.
Before, the animations would occur for all zombies whenever player was
in range of a single zombie. Now, it only occurs for the zombie that the
player is in range of.