Implemented random zombie spawns; Killed zombies dont respawn.
What this means is that when a zombie gets killed and a player enters +
exits a house, the number of zombies remaining are stored and saved, so
that all of the zombies that you killed don't respawn again.

Also, the zombies spawn at different locations after the game restarts
or the player enters and exits the house.